Module type Requester.PRECHECK

When a requested value is received, it goes to a validation phase.

At fetching time, the client gives a param. At notification time, the client provides a notified_value. precheck tries to construct a value from param and notified_value.

If no validation is needed. precheck is defined as

let precheck k () v -> Some v

type key

type param

type notified_value

type value

val precheck : key -> param -> notified_value -> value option